+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+* Add `Signal#tap` method
+
 ## 6.0.0 (2019-03-23)
 
 * Refactor combinators to fix cyclical dependency warnings